+++ src/models/iam/role.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, Integer, String, UniqueConstraint
+from sqlalchemy import BigInteger, Boolean, Column, String, UniqueConstraint
 from sqlalchemy.orm import relationship
 
 from src.models.base import BaseModel, RemarkMixin, SoftDeleteMixin, TimestampMixin
@@ -14,7 +14,7 @@ class Role(BaseModel, TimestampMixin, SoftDeleteMixin, RemarkMixin):
     )
 
     name = Column(String(20), nullable=False, comment="角色名称")
-    tenant_id = Column(Integer, nullable=True, comment="租户ID（null=系统级）")
+    tenant_id = Column(BigInteger, nullable=True, comment="租户ID（null=系统级）")
     is_system = Column(Boolean, default=False, nullable=False, comment="系统内置角色，创建后不可修改")
 
     users = relationship("User", secondary=user_role_association, back_populates="roles")
